+++ crates/objc2-encode/src/lib.rs
@@ -117,3 +117,4 @@ mod static_str;
 pub use self::encode::{Encode, EncodeArguments, EncodeConvert, RefEncode};
 pub use self::encoding::Encoding;
 pub use self::encoding_box::EncodingBox;
+pub use self::parse::ParseError;